[init] Initial commit of container-reposync
The Dockerfile and scripts in this repo create a container image
that is able to sync content from both the Unbreakable Linux Network
and the Oracle Linux yum server with repo URLs that match those
found on https://yum.oracle.com.

It also includes a web server image that can be used to serve the content.

Signed-off-by: Avi Miller <avi.miller@oracle.com>